+++ test/jest/__mock__/stripesCore.mock.js
@@ -1,0 +1,6 @@
+import React from 'react';
+
+jest.mock('@folio/stripes/core', () => ({
+  stripesShape: {},
+  IfPermission: jest.fn(({ children }) => <div>{children}</div>),
+}));